Added and implemented custom lexer
Implemented my own custom lexer + derive macro called "pglex".
Unfortunately this is necessary as Logos currently has some breaking
bugs and development does not seem very active, unfortunately.
Still needs testing and is probably VERY slow.

+++ pgs/src/parser/ast.rs
@@ -8,7 +8,7 @@ use std::{
 #[derive(PartialEq, Debug, Clone)]
 pub enum Expression {
     IntLiteral(i64),
-    FloatLiteral(f64),
+    FloatLiteral(f32),
     StringLiteral(String),
     BoolLiteral(bool),
     Variable(String),
